requirements.txt
aiofile==3.7.4
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.1
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.2
cachetools==5.0.0
caio==0.9.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
distlib==0.3.4
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.76.0
filelock==3.7.0
google-auth==2.6.6
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyhumps==3.5.3
pyparsing==3.0.8
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-env==0.6.2
pytest-timeout==2.1.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.18.0
toml==0.10.2
tomli==2.0.1
tox==3.25.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.1
